remove HomeListMain.vue
